reintegrate fred2

+++ src/fred2/campaigneditordlg.cpp
@@ -485,7 +485,7 @@ void campaign_editor::OnSelchangedSexpTree(NMHDR* pNMHDR, LRESULT* pResult)
 	}
 
 	// get identifier of parent
-	node = m_tree.GetItemData(h);
+	node = static_cast<int>(m_tree.GetItemData(h));
 	for (i=0; i<Total_links; i++){
 		if ((Links[i].from == Cur_campaign_mission) && (Links[i].node == node)){
 			break;
