+++ src/fred2/shipeditordlg.cpp
@@ -1372,7 +1372,7 @@ int CShipEditorDlg::update_ship(int ship)
 	// m_persona holds the index into the list.  Get the item data associated with this index and then
 	// assign to the ship taking care that we check for the NO_PERSONA_INDEX id
 	box = (CComboBox *)GetDlgItem(IDC_SHIP_PERSONA);
-	persona = box->GetItemData(m_persona);
+	persona = static_cast<int>(box->GetItemData(m_persona));
 	if ( persona == NO_PERSONA_INDEX )
 		persona = -1;
 
@@ -1397,7 +1397,7 @@ int CShipEditorDlg::update_ship(int ship)
 		m_arrival_delay.save(&Ships[ship].arrival_delay);
 		m_departure_delay.save(&Ships[ship].departure_delay);
 		if (m_arrival_target >= 0) {
-			z = ((CComboBox *) GetDlgItem(IDC_ARRIVAL_TARGET)) -> GetItemData(m_arrival_target);
+			z = static_cast<int>(((CComboBox *) GetDlgItem(IDC_ARRIVAL_TARGET)) -> GetItemData(m_arrival_target));
 			MODIFY(Ships[ship].arrival_anchor, z);
 
 			// if the arrival is not hyperspace or docking bay -- force arrival distance to be
@@ -1419,7 +1419,7 @@ int CShipEditorDlg::update_ship(int ship)
 				}
 			}
 		}
-		z = ((CComboBox *)GetDlgItem(IDC_DEPARTURE_TARGET))->GetItemData(m_departure_target);
+		z = static_cast<int>(((CComboBox *)GetDlgItem(IDC_DEPARTURE_TARGET))->GetItemData(m_departure_target));
 		MODIFY(Ships[ship].departure_anchor, z );
 	}
